refactor: findById 대신 existsById 사용으로 성능 최적화

+++ src/main/java/com/process/clash/application/shop/recommendedproduct/service/DeleteRecommendedProductService.java
@@ -22,8 +22,9 @@ public class DeleteRecommendedProductService implements DeleteRecommendedProduct
 
         checkAdminPolicy.check(command.actor());
 
-        recommendedProductRepositoryPort.findById(command.recommendedProductId())
-                .orElseThrow(RecommendedProductNotFoundException::new);
+        if (!recommendedProductRepositoryPort.existsById(command.recommendedProductId())) {
+            throw new RecommendedProductNotFoundException();
+        }
 
         recommendedProductRepositoryPort.deleteById(command.recommendedProductId());
     }
